+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Updated the default value of AllowBlobPublicAccess and AllowCrossTenantReplication to false when creating a storage account. More details: https://techcommunity.microsoft.com/t5/azure-storage-blog/azure-storage-updating-some-default-security-settings-on-new/ba-p/3819554
+    - `New-AzStorageAccount`
 
 ## Version 6.0.1
 * Updated error message when storage context is missing in a cmdlet input